// src/domain/enums/ExerciseCategory.ts
/**
 * Training category an exercise belongs to.
 */
export enum ExerciseCategory {
  Cardio = 'CARDIO',
  OlympicWeightlifting = 'OLYMPIC_WEIGHTLIFTING',
  Plyometrics = 'PLYOMETRICS',
  Powerlifting = 'POWERLIFTING',
  Strength = 'STRENGTH',
  Stretching = 'STRETCHING',
  Strongman = 'STRONGMAN',
}